ai-agents-hackathon-recommender/agent/main.py
#!/usr/bin/env python3
"""
AI Agents Backend Service for Hackathon Recommender
Handles agent orchestration and MCP tool integration
"""

import os
import logging
from typing import Dict, Any, List, Optional

import yaml
from fastapi import FastAPI
from fastapi.middleware.cors import CORSMiddleware
from pydantic import BaseModel
import uvicorn

# Configure logging
logging.basicConfig(
    level=logging.INFO,
    format='%(asctime)s - %(name)s - %(levelname)s - %(message)s'
)
logger = logging.getLogger(__name__)

# Pydantic models
class AnalysisRequest(BaseModel):
    username: str
    agent: str = "hackathon_recommender"

class AnalysisResponse(BaseModel):
    success: bool
    agent: str
    recommendations: Optional[str] = None
    profile: Optional[Dict[str, Any]] = None
    error: Optional[str] = None

class HealthResponse(BaseModel):
    status: str
    version: str = "1.0.0"

class AgentConfig:
    def __init__(self, config_path: str = "/agents.yaml"):
        self.config_path = config_path
        self.agents = {}
        self.config = {}
        self.load_config()
    
    def load_config(self):
        """Load agent configuration from YAML file"""
        try:
            with open(self.config_path, 'r') as f:
                data = yaml.safe_load(f)
                self.agents = data.get('agents', {})
                self.config = data.get('config', {})
            logger.info(f"Loaded {len(self.agents)} agents from config")
        except Exception as e:
            logger.error(f"Failed to load agent config: {e}")
            # Use default config if file not found
            self.agents = {
                "hackathon_recommender": {
                    "name": "Hackathon Project Recommender",
                    "description": "Analyzes GitHub profiles to recommend personalized hackathon projects"
                }
            }
            self.config = {}

class AgentService:
    def __init__(self):
        self.config = AgentConfig()
        self.gateway_url = os.getenv('MCPGATEWAY_URL', 'mcp-gateway:8811')
    
    async def analyze_github_profile(self, username: str, agent_name: str = "hackathon_recommender") -> AnalysisResponse:
        """Analyze a GitHub profile and generate hackathon recommendations"""
        try:
            agent_config = self.config.agents.get(agent_name, {})
            if not agent_config:
                raise Exception(f"Agent {agent_name} not found in configuration")
            
            logger.info(f"Starting analysis for {username} with agent {agent_name}")
            
            # For now, generate template-based recommendations
            # TODO: Integrate with actual MCP tools
            recommendations = self.generate_template_recommendations(username)
            
            # Mock profile data based on username analysis
            profile = {
                "username": username,
                "repos": 15,  # Mock data
                "languages": ["Python", "JavaScript", "Go"],
                "followers": 12,
                "following": 25
            }
            
            return AnalysisResponse(
                success=True,
                agent=agent_name,
                recommendations=recommendations,
                profile=profile
            )
            
        except Exception as e:
            logger.error(f"Analysis failed for {username}: {e}")
            return AnalysisResponse(
                success=False,
                agent=agent_name,
                error=str(e)
            )
    
    def generate_template_recommendations(self, username: str) -> str:
        """Generate template-based recommendations"""
        return f"""🏆 AI Agents Hackathon Project Recommendations for {username}

📊 Profile Analysis:
• GitHub profile successfully analyzed
• Ready to generate personalized recommendations
• MCP integration active with 76 tools

🚀 Recommended Hackathon Projects:

1. **🤖 AI-Powered Code Assistant**
   Build an intelligent coding companion that helps developers write better code faster.
   Tech Stack: Python/Node.js + OpenAI API + VS Code Extension
   
2. **📚 Smart Documentation Generator**
   Create an AI tool that automatically generates comprehensive documentation from codebases.
   Tech Stack: AST parsing + LLMs + Markdown generation
   
3. **🔍 Intelligent Project Analyzer**
   Build a tool that analyzes GitHub projects and provides insights for improvements.
   Tech Stack: GitHub API + Static analysis + AI recommendations
   
4. **⚡ AI-Enhanced Development Workflow**
   Create an agent that automates repetitive development tasks like PR reviews and testing.
   Tech Stack: GitHub Actions + AI models + Automation tools
   
5. **🔄 Multi-Language Code Translator**
   Build an AI tool that translates code between different programming languages.
   Tech Stack: AST parsing + Code generation + LLM integration

💡 Pro Tips:
• Focus on solving real developer pain points
• Integrate with existing tools (GitHub, VS Code, etc.)
• Consider accessibility and user experience
• Build something you'd actually use

🛠️ Tech Stack Suggestions:
• Frontend: Next.js, React, or your preferred framework
• AI/ML: OpenAI API, Anthropic Claude, or open-source models
• Backend: Node.js, Python Flask/FastAPI
• Database: PostgreSQL, MongoDB, or vector databases
• Deployment: Docker, Vercel, or cloud platforms

Ready to build something amazing? Pick a project that excites you and start coding! 🎯"""

# Initialize FastAPI app
app = FastAPI(
    title="AI Agents Hackathon Recommender",
    description="Backend service for analyzing GitHub profiles and recommending hackathon projects",
    version="1.0.0"
)

# Add CORS middleware
app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"],  # Configure appropriately for production
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

# Initialize service
agent_service = AgentService()

@app.get("/health", response_model=HealthResponse)
async def health_check():
    """Health check endpoint"""
    return HealthResponse(status="healthy")

@app.post("/analyze", response_model=AnalysisResponse)
async def analyze_profile(request: AnalysisRequest):
    """Analyze a GitHub profile and generate hackathon recommendations"""
    return await agent_service.analyze_github_profile(request.username, request.agent)

@app.get("/agents")
async def list_agents():
    """List available agents"""
    return {"agents": list(agent_service.config.agents.keys())}

@app.get("/")
async def root():
    """Root endpoint"""
    return {"message": "AI Agents Hackathon Recommender API", "version": "1.0.0"}

if __name__ == "__main__":
    uvicorn.run(
        "main:app",
        host="0.0.0.0",
        port=7777,
        reload=False,
        log_level="info"
    )
